feat(cpProgress): extract progress logic
