+++ src/utils/campaigns/getWidgetCampaignProgress/index.ts
@@ -1,82 +1,14 @@
-import * as db from "@src/features/db";
-
-interface bugsByUseCaseQueryResults {
-  total: number;
-  id: number;
-  title: string;
-  content: string;
-  simple_title: string;
-  info: string;
-  prefix: string;
-}
+import { formatUseCaseProgress } from "./formatProgress";
+import { getUseCaseProgress } from "./getUseCaseProgress";
 
 // UC allowed responses 12.5 | 37.5 | 62.5 | 87.5
-
-const UC_PROGRESS_1 = { max: 25, value: 12.5 as const }; //  0%  - 24.99%
-const UC_PROGRESS_2 = { max: 50, value: 37.5 as const }; //  25% - 49.99%
-const UC_PROGRESS_3 = { max: 75, value: 62.5 as const }; //  50% - 74.99%
-const UC_PROGRESS_4 = { max: 100, value: 87.5 as const }; // 75% - 100%
-
 const getWidgetUseCaseProgress = async (
   cp: StoplightComponents["schemas"]["CampaignWithOutput"]
 ): Promise<12.5 | 37.5 | 62.5 | 87.5 | 100> => {
-  // Selected testers
-  const testersQuery = `SELECT 
-    p.id, 
-    c.group_id
-    FROM wp_crowd_appq_has_candidate c
-           JOIN wp_appq_evd_profile p ON (p.wp_user_id = c.user_id)
-    WHERE c.campaign_id = ?
-    AND accepted = 1 GROUP BY p.id`;
+  const { groups, usecases } = await getUseCaseProgress(cp.id);
 
-  const testers = await db.query(db.format(testersQuery, [cp.id]));
-
-  if (!testers.length) {
-    return UC_PROGRESS_1.value;
-  }
-
-  const defaultGroups = {
-    0: 0, // All
-    1: 0, // Group 1
-    2: 0, // Group 2
-    3: 0, // Group 3
-    4: 0, // Group 4
-  };
-
-  // reduce tester to group
-  const groups = testers.reduce(
-    (
-      acc: { [x: number]: number },
-      tester: { id: number; group_id: number }
-    ) => {
-      acc[0] += 1; // Every tester is in the all group
-
-      if (!acc[tester.group_id]) {
-        acc[tester.group_id] = 1;
-      } else {
-        acc[tester.group_id] += 1;
-      }
-      return acc;
-    },
-    defaultGroups
-  );
-
-  // Use cases status
-  const ucQuery = `SELECT 
-    uc.id, 
-    g.group_id, 
-    uc.campaign_id, 
-    COUNT(DISTINCT t.tester_id) as completions
-    FROM wp_appq_campaign_task uc
-      LEFT JOIN wp_appq_user_task t ON (t.task_id = uc.id AND t.is_completed = 1)
-      JOIN wp_appq_campaign_task_group g ON (g.task_id = uc.id)
-    WHERE uc.campaign_id = ?
-    group by uc.id`;
-
-  const uc = await db.query(db.format(ucQuery, [cp.id]));
-
-  if (!uc.length) {
-    return UC_PROGRESS_1.value;
+  if (!usecases.length || !groups[0]) {
+    return formatUseCaseProgress();
   }
 
   const completion = {
@@ -84,20 +16,17 @@ const getWidgetUseCaseProgress = async (
     actual: 0,
   };
 
-  uc.forEach((useCase: { group_id: number; completions: number }) => {
+  usecases.forEach((useCase: { group_id: number; completions: number }) => {
     completion.expected += groups[useCase.group_id];
     completion.actual += useCase.completions;
   });
 
   if (completion.expected === 0 || completion.actual === 0)
-    return UC_PROGRESS_1.value;
+    return formatUseCaseProgress();
 
   const progress = (completion.actual / completion.expected) * 100;
 
-  if (progress < UC_PROGRESS_1.max) return UC_PROGRESS_1.value;
-  else if (progress < UC_PROGRESS_2.max) return UC_PROGRESS_2.value;
-  else if (progress < UC_PROGRESS_3.max) return UC_PROGRESS_3.value;
-  else return UC_PROGRESS_4.value;
+  return formatUseCaseProgress(progress);
 };
 
 export const getWidgetCampaignProgress = async (
